method to flag + verify dependent params, more...
With EasyFabric, some parameters are not mandatory in some cases, but become mandatory in others.  We've added a method, build_mandatory_params(), that allows such parameters to be specified, and another method, validate_dependencies(), that verifies these have been either set by the user, or that there's a default value for them.

Other changes:

1. add a couple more fabric parameters
- anycast_lb_id
- anycast_rp_ip_range

2. rename a couple vars that I didn't understand the meaning of e.g. inv_spec renamed to params_spec

3. shorten the names of a few vars to make it easier to stay under a 79 character line length.  e.g. _fabric_nv_pairs_default to _default_nv_pairs

4. translate_to_ndfc_nv_pairs() Handle dunder and camelCase key