dict and accessor method for aliased parameter values
These are used to present more user-friendly messages in cases where the value is the playbook is translated to something else prior to sending to NDFC.  We always want to use the playbook values in user messages rather than values sent to NDFC (if these differ).

Also, change "any" to "__any__" in self._mandatory_params to avoid any future conflict with a valid value "any" (so far, there are no known instances of this, so we're just being extra-careful).